add markers for store location and add detail for each store

+++ MobileApp/lib/models/Store.dart
@@ -31,41 +31,46 @@ class Store {
   String email;
   String phone;
   String address;
-  int longitude;
-  int latitude;
+  double longitude;
+  double latitude;
   int locationId;
   bool isActive;
   Location? location;
   List<dynamic> storeSalesDetails;
   List<dynamic> users;
 
-  factory Store.fromJson(Map<String, dynamic> json) => Store(
-    id: json["id"],
-    name: json["name"],
-    email: json["email"],
-    phone: json["phone"],
-    address: json["address"],
-    longitude: json["longitude"],
-    latitude: json["latitude"],
-    locationId: json["locationId"],
-    isActive: json["isActive"],
-    location: json["location"] == null  ? null : Location.fromJson(json["location"]) ,
-    storeSalesDetails: List<dynamic>.from(json["storeSalesDetails"].map((x) => x)),
-    users: List<dynamic>.from(json["users"].map((x) => x)),
-  );
+  factory Store.fromJson(Map<String, dynamic> json) =>
+      Store(
+        id: json["id"],
+        name: json["name"],
+        email: json["email"],
+        phone: json["phone"],
+        address: json["address"],
+        longitude: json["longitude"].toDouble(),
+        latitude: json["latitude"].toDouble(),
+        locationId: json["locationId"],
+        isActive: json["isActive"],
+        location: json["location"] == null ? null : Location.fromJson(
+            json["location"]),
+        storeSalesDetails: List<dynamic>.from(
+            json["storeSalesDetails"].map((x) => x)),
+        users: List<dynamic>.from(json["users"].map((x) => x)),
+      );
 
-  Map<String, dynamic> toJson() => {
-    "id": id,
-    "name": name,
-    "email": email,
-    "phone": phone,
-    "address": address,
-    "longitude": longitude,
-    "latitude": latitude,
-    "locationId": locationId,
-    "isActive": isActive,
-    "location": location?.toJson(),
-    "storeSalesDetails": List<dynamic>.from(storeSalesDetails.map((x) => x)),
-    "users": List<dynamic>.from(users.map((x) => x)),
-  };
+  Map<String, dynamic> toJson() =>
+      {
+        "id": id,
+        "name": name,
+        "email": email,
+        "phone": phone,
+        "address": address,
+        "longitude": longitude,
+        "latitude": latitude,
+        "locationId": locationId,
+        "isActive": isActive,
+        "location": location?.toJson(),
+        "storeSalesDetails": List<dynamic>.from(
+            storeSalesDetails.map((x) => x)),
+        "users": List<dynamic>.from(users.map((x) => x)),
+      };
 }
